Add an option to use a custom file system for build artifacts.
Add an option for `BlazeModule`s to be able to register a custom file system
implementation to be used for build artifacts (e.g. sources, outputs,
`BUILD/WORKSPACE/.bzl` files). This file system will be used by
`PackageLocator` and all `ArtifactRoot`s.

Add a mechanism to propagate `WorkspaceInfoFromDiff` during workspace
synchronization. `DiffAwareness` can emit information about the workspace if it
is available. Add a mechanism to allow listeners to subscribe to that
information when (and if) it is returned from the diff.

PiperOrigin-RevId: 374979680