Add --deduplicate_depsets to aquery.
This flag would instruct aquery to remove duplicate subsets in dep_set_of_files
before writing the proto/textproto/json output. The duplication caused the
aquery output to be corrupt in some cases, an issue which we haven't managed to
get to the bottom of yet. Duplicate subsets carry little meaning anyway, since
in the end we mostly care about the effective collection of inputs of actions.
It's already not possible to create NestedSets with duplicate subsets using
NestedSetBuilder.

The deduplication is done with a flag in case we want to study the depsets with
duplicate subsets, which is really the unfortunate exception rather than the
norm. The flag is set to True by default.

PiperOrigin-RevId: 375107739