Do not attempt to expand omitted tree artifacts in build events.
When constructing the build events, we will add tree artifact expansions to the
`named_set_of_files`. In case of the tree artifact is orphaned, we will not
have the expansions available for it, resulting with a
[NullPointerException](https://github.com/bazelbuild/bazel/blob/300c11ef5eb00978cbaa4d3a14ccfee10d9c87f8/src/main/java/com/google/devtools/build/lib/actions/CompletionContext.java#L124).

Do not attempt to expand tree artifacts for the events when they are omitted.

PiperOrigin-RevId: 376052463