Remove the boolean field `AbstractQueueVisitor#usingPriorityQueue`.
It's currently always false, and there's no robust way to fix it because `AbstractQueueVisitor` allows passing a custom `ExecutorService`, which does not provide a way to determine the type of the underlying work queue.

This boolean was only used for a precondition check that the `Runnable` being submitted is also a `Comparable`. However, the priority queue implementation throws `ClassCastException` anyway if it needs to make a comparison and cannot.

PiperOrigin-RevId: 376829066