Enqueue children of restarted nodes at a higher priority and increase priority each time a node is restarted.
Instead of using a "regular" priority from `globalEnqueuedIndex`, children of restarted nodes jump to the top of the priority queue. In wide graphs with many restarted nodes, the queue may be saturated with maximum priority evaluations,
preventing the children of these nodes from getting evaluated. This is especially bad during execution, where the `stateMap` in `ActionExecutionFunction` retains input mappings for pending (restarted but not completed) actions.

PiperOrigin-RevId: 376881478