Use `aapt2 convert` instead of relinking final APK as static lib
The previous code to convert a proto APK to binary involved
re-interpreting it as a static library, and relinking into a new final APK.
Static libraries do not normally have IDs assigned, which was worked around by
passing the assigned IDs as `--stable-ids` when relinking.

That worked as aapt2 does not verify that static libraries do not have IDs
assigned, but it *does* verify that they have the default package (0x7f),
preventing the use of `package-id`.

Aapt2 provides the `convert` command, which should obviate the need for
relinking.

This includes two workarounds to preserve existing behavior:

1. APKs in `assets/` are *always* left uncompressed.
2. Recompress XML files in `res/` after `aapt2 optimize`. AAPT2 looks for
incompressible files and decompresses them, presumably to avoid unnecessary
overhead. This takes place before conversion of XML from proto to binary, when
the compressibility generally improves (by several kB in sampled APKs).

PiperOrigin-RevId: 377047012