Remote: Report checking cache status before the action is scheduled to run remotely.
Add a Caching state to ActionState. This state indicates the action is checking the cache and should be happened before Scheduling state.

Change ProgressStatus from enum to interface so that we can pass more data to the event handler (which is required to report upload/download details later).

Fixes https://github.com/bazelbuild/bazel/issues/13531.