Delete accidentally-abandoned DocumentationFailureTests
The test itself is broken (no exception is thrown) but not running. AFAICT it's
been this way since it was checked in. The issue is the test class isn't under
com.google.devtools.build.*lib* where the AllTests runner is looking. While the
AllTests runner isn't _supposed_ to support empty test suites, it has a bug
where it +/- unconditionally picks up the empty dummy test it uses to support
the "no tests are ok" mode.

PiperOrigin-RevId: 377563200