Move some OOM handling (including heap dumps) from `RetainedHeapLimiter` to `BugReport`, and output the heap dump under the output base.
This way all OOMs are handled exactly the same, regardless of where they come from. Even standard JVM OOMs are caught and channeled through `BugReport`, so `-XX:+HeapDumpOnOutOfMemoryError` has no effect (unless the server were to somehow OOM outside the scope of a command, but it didn't seem worthwhile to leave the JVM arg for that).

This also has the side effect of using the command's `Reporter` for async crashes thanks to `BlazeRuntime#fillInCrashContext`, resulting in colored output.

PiperOrigin-RevId: 377625413