Delete defunct test suite infrastructure
Everything but supportedOs is never read in practice. They were probably used
at some point, but we've migrated much of the test suite infrastructure over to
regular old BUILD-defined tests. Deleted the attributes that were easy, but left
size since it'll be a larger effort to clean up all of its vestigial uses.

`tolerateEmptyTestSuites` is no longer used. In fact it was actually harmful
since it gave the impression that it actually worked - `TautologyTest` would get
picked up by the classpath-scanning test search of c.g.devtools.build.lib, which
every use of the test suite builder would use, so test suites were never really
empty in practice and would blindly succeed due to `TautologyTest` blindly
succeeding.

Alternatively we _could_ go the other direction and leave things as they are
with the eventual goal of improving the test-suite infra. However, we've already
moved quite a bit away from the point where it'd make sense to do that, which is
why it was so easy do delete all of this functionality. Also, as noted in the
previous paragraph, the framework actually hides bugs, so removing it reduces
room for error.

PiperOrigin-RevId: 377972815