BEP events retain ActionInputMap holding only important artifacts.
To support IDE use-cases, https://github.com/bazelbuild/bazel/commit/06c956623fdb81465f884c170df48bf38b59cadb changed target completion events to keep
a reference to the ActionInputMap in order to keep around metadata specifying if
an output file is a directory or not. This worked but caused an unexpected
increase in heap high-watermark, causing OOMs.

The unexpected heap usage was traced to large _validation output groups holding
tens of thousands of artifacts per target. This output group is not included in
BEP so the original change was in fact retaining far more memory than necessary
to implement the intended optimization.

With this change, a second ActionInputMap is prepared which includes only the
"important" artifacts and is sized appropriately. Important artifacts are those
which are not produced by internal output groups and are exactly the artifacts
whose URIs are written to BEP. Maintaining this second ActionInputMap costs
additional latency in CompletionFunction proportional to the number of
important artifacts, though care is taken to avoid doubling the runtime of
CompletionFunction in the case where all artifacts are important.

PiperOrigin-RevId: 378793974