Preserve `PathFragment` passed to `ActionInputHelper::fromPath(PathFragment)`.
`ActionInputHelper::fromPath` allows to create an action input from an exec path or a string representing it. In case of `PathFragment`, we will get the path string of it and recreate a new `PathFragment` in `getExecPath`. Store the `PathFragment` instead. Change callers which already have a `PathFragment` to use the override accepting it.

PiperOrigin-RevId: 378907363