Remove check against tree artifacts from `ParameterFileWriteAction`.
`ParameterFileWriteAction::getArguments` will fail if called on a parameter file which has tree artifacts in the command line. This is possible when running `aquery` and will crash the Bazel in such case. Remove the check to avoid crashing the server.

Please note that aquery will return incorrect contents for the param file since it will have unexpanded tree artifact. At time of aquery, we do not have tree artifact expansions available, hence this is the best we can do. This behavior is consistent with aquery results for action arguments using `Args::add_all` with tree artifact expansion.

PiperOrigin-RevId: 379622158