Simplify and enable outputfilter tests
For OutputFilter a whole bunch of fake java code, tweaking flags, and
reconstructing error messages seems like a bit of overkill to test an
interface that just converts a string to a boolean. Remove all of that and
focus on the basic functionality - we have shell tests elsewhere that should
verify enough of the end-to-end functionality.

For AutoOutputFilter creating BUILD files was superfluous and the only thing
pulling in BuildIntegrationTestCase, so remove it. I avoided going too deep
on that test, just did bare minimum to cut BuildIntegrationTestCase out.

PiperOrigin-RevId: 379723974