Make transition directory name fragment reflect output directory name
If output directory name is overriden, the output path does no longer
consists of the transition directory name fragment.

The calculation of transition directory name fragment does not reflect
this situation and instead we can get conflicts due to this name itself.

This patch makes sure that a the transition directory name fragment
gets a unique hash for a given output directory name.