Flip default for experimental_forward_instrumented_files_info_by_default
This changes the default behavior for coverage from "forward nothing" to "forward InstrumentedFilesInfo from non-tool dependencies". This is still not quite ideal behavior, since typically attributes which provide source files (like srcs) are non-tool dependencies. But it's an improvement because it avoids the need for explicit coverage configuration for the entire chain of dependencies from test to source file.

In cases where an aspect returned InstrumentedFilesInfo previously, the InstrumentedFilesInfo provider for the base rule is ignored, so the behavior will be unchanged.

The bit about coverage_common.instrumented_files_info is merged into the related section under "Advanced Topics" in the rules documentation, since it's no longer required for the whole dependency chain to configure coverage explicitly in order for that to work at all.

RELNOTES: Forward coverage-instrumented files from non-tool dependencies by default.
PiperOrigin-RevId: 380224680