Include the Starlark stack trace as part of the crash's stack trace for all unhandled throwables originating in Starlark evaluation.
In keeping with tradition of not wrapping `Error` in `AbstractParallelEvaluator`, different exception types are used for `RuntimeException` and `Error`.

Additionally embed a context supplier in `StarlarkThread`. This is specifically targeted for the case of a crash during rule analysis, where the call stack itself doesn't include enough information to identify the rule (since the `BUILD` file is evaluated during loading).

PiperOrigin-RevId: 380246681