Make sure correct libraries get linked to cc_test
This is done in the native objc_library implementation by re-wrapping each
library_to_link in a new one that doesn't have a .so. This has an additional
memory cost, to avoid a regression in the Starlark implementation we stop doing
this re-wrapping. The right library is chosen now when it actually matters
during linking.

RELNOTES:none
PiperOrigin-RevId: 381025005