Reduce garbage in header discovery.
The `ImmutableMap` copy creates significant garbage in some builds. Since `HeaderDiscovery` instances are only constructed for one-time use in practice, make `discoverInputsFromDependencies` a static method.

The set of tree artifacts is dropped in favor of just a map because the `NestedSet` from which they originate will already deduplicate. Additionally, it is not necessary to also store them in the main derived artifact map.

PiperOrigin-RevId: 381027740