Skip the call to `type.cast(value)` in `AbstractAttributeMapper#get`.
The call to `cast` shows up as taking significant cpu in profiles. My guess is that it's from virtual method overhead, but I have not been able to reproduce. Since any type mismatches are caught during rule construction, just use an unchecked cast.

To keep some level of type safety, explicitly check for `SelectorList`, which has been a source of bugs in the past (b/19867975).

I can follow up with more principled changes to reduce the overall number of calls to `get`, but I'm curious whether this alone will have an effect.

PiperOrigin-RevId: 381312360