Improve creation of `Label` and `RepositoryName` instances.
In `Label#getRelativeWithRemapping`, when the current label is in the main repository, default the new label to the main repo as well. This saves the intermediate step of creating (and interning) a garbage `Label` with the default repository, only for `resolveRepositoryRelative` to subsequently create another `Label` in the main repo.

In `RepositoryName#create`, optimize for the common repository names (default and main) by skipping the interner cache and other unnecessary checks.

PiperOrigin-RevId: 381857421