Remove AggregatingMiddleman from toolchains.
After flipping --experimental_enable_aggregating_middleman to false, we can remove the creation & usage of AggregatingMiddleman in the cc toolchain and other parts of the code base that makes use of MiddlemanProvider to add inputs to actions.

The next step is to remove the special handling for AggregatingMiddleman in Skyframe itself.

PiperOrigin-RevId: 381861891