Improve Starlark transition efficiency.
Specifically, avoid unnecessarily cloning BuildOptions - a known expensive operation. This is particularly important for the common pattern of transitions that are often noops. For example, a transition of the form "change the CPU but only if the CPU hasn't already been changed by a parent target".

This change avoids the expensive cloning for noops.

Benchmark tests on a large target with ~500,000 Starlark transitions - the vast majority of which are noops - show a 16% reduction in build analysis time.

Thanks to https://github.com/oquenchil for the first draft of this code.

PiperOrigin-RevId: 382351589