Allow ComputedDefault to declare if resolvable w/o configuration.
TargetSerializer has an interest in serializing the results of a ComputedDefault; however, this requires the ComputedDefault be resolvable at loading time, without a configuration.

Roughly, this means .get is never called on a configurable attribute.

PS: Ideally, this would be implemented slightly differently. Perhaps by creating a subclass of ComputedDefault exposing a resolve method that explicitly takes a RawAttributeMapper. (The typing of AttributeMapper is also a bit tricky. RawAttributeMapper likely should not even have a get method.) However, cleaning this up is left as a future exercise.
PiperOrigin-RevId: 382392871