Some clarity cleanups in NestedSet
We use some rather novel tricks for efficiency, try to make it easier to follow.

Documentation is left mostly the same except switched to javadoc so that one
doesn't need to scroll all the way to the top of the file to get context on what
the purpose of something is. I removed the explicit mention of depth in
memo/NO_MEMO since it was inconsistent and I was having a hard time figuring out
who the winner should be.

Changed up some ordering to better achieve java customs / not mix up static and
instance state.

Got more verbose with some variable names. `bytes` in particular took me a while
to parse - it's the offset into memo, not the number of bytes in any thing in
the immediate area (technically maybe how many bytes pos translates to, but a
bit of a leap).

Switched the resize-memo check to use a positive condition - I think it reads
more naturally and negates the need for some of the larger comments.

Use `pos++` instead of `++pos` for consistency within the file, and because
we've run into some tricky unexplainable exceptions that seem to just go away
when using the postfix ++.

PiperOrigin-RevId: 382609974