Remote: Display download progress when actions are downloading outputs from remote cache.
Normally, when executing action with remote cache/execution, the UI only display the "remote"/"remote-cache" strategy:
```
[500 / 1000] 500 actions, 3 running
    [Sched] Executing genrule //:test-1;
    Executing genrule //:test-2; 2s remote
    Executing genrule //:test-3; 3s remote ...
```

However, it doesn't tell users what is happening under the hood. #13555 fixed the confusion which the UI display the action is scheduling while it is actually downloading the outputs.

With this change, Bazel will display the downloads if action is downloading outputs. e.g.
```
[500 / 1000] 500 actions, 3 running
    [Sched] Executing genrule //:test-1; 1s remote
    Executing genrule //:test-2; Downloading 2.out, 20.1 KiB / 100 KiB; 2s remote
    Executing genrule //:test-3; 3s remote ...
```

Add a generic `ActionProgressEvent` which can be reported within action execution to display detailed execution progress for that action.

Closes #13557.

PiperOrigin-RevId: 383224334