Use AspectsListBuilder to build top level aspects lists
This CL uses the newly added class `AspectsListBuilder` to build the list of top level aspects with their details. This is a step towards using enabling `requires` attribute for top level aspects.

Since `AspectsListBuilder` does not allow duplicates in the aspects list, some of the tests needed to be updated to follow that.

PiperOrigin-RevId: 383363263