Optimize DefaultInfo constructed from native rules
When rules implemented in starlark inspect DefaultInfo in their dependencies (ctx.attr.deps[x][DefaultInfo] or ctx.attr.deps[x].files) and those dependencies are in native, DefaultInfo gets constructed (with 5 fields).

This cl optimises it and returns DelegatingDefaultInfo implementation with 1 field.

There is a rare case where a Starlark target can return DelegatingDefaultInfo (obtaining it from native rule) - in this case the typecasts in StarlarkRuleConfiguredTargetUtil needed to be fixed.

Further optimisations are possible - data runfiles are deprecated and most Starlark rules don't use it - so another DefaultInfo with only 3 fields could be constructed. Or data runfiles could be cleaned up completely.

Also FilesProvider should be cleaned up in favour of DefaultInfo. It can't be constructed in Starlark, and eventually when things are rewritten in Starlark it will disappear.

PiperOrigin-RevId: 384184094