Resolve symlinks to clang in Unix toolchain
Fully resolves symlinks to the CC binary in the auto-configured Unix
toolchain if CC is clang.

Background:
clang does not find the libc++ headers when invoked through a symlink in
a different directory with `-no-canonical-prefixes`. On Ubuntu 21.04
with clang-12 installed:

```shell
$ /usr/bin/clang -E -xc++ - -v -stdlib=libc++
 /usr/lib/llvm-12/bin/../include/c++/v1
 /usr/local/include
 /usr/lib/llvm-12/lib/clang/12.0.0/include
 /usr/include/x86_64-linux-gnu
 /usr/include
End of search list.

$ /usr/bin/clang -E -xc++ - -v -stdlib=libc++ -no-canonical-prefixes
 /usr/local/include
 /usr/lib/clang/12.0.0/include
 /usr/include/x86_64-linux-gnu
 /usr/include
End of search list

$ readlink -f /usr/bin/clang
/usr/lib/llvm-12/bin/clang

$ /usr/lib/llvm-12/bin/clang -E -xc++ - -v -stdlib=libc++ -no-canonical-prefixes
 /usr/lib/llvm-12/bin/../include/c++/v1
 /usr/local/include
 /usr/lib/llvm-12/lib/clang/12.0.0/include
 /usr/include/x86_64-linux-gnu
 /usr/include
End of search list
```

Both the first and the last command find all headers files, the second
one representing the current Bazel behavior does not.