Do not flatten tree artifacts in `ActionInputMap`.
When constructing an `ActionInputMap`, we will add all of the action inputs to it, hashed on the exec path. For tree artifacts, we would add both the tree and all of the children from its expansion. That poses a memory usage problem in presence of large tree artifacts -- for one such tree artifact, we would create `O(actions depending on it)` references to each of the files in the tree artifacts.

Reduce the memory usage of `ActionInputMap` by storing a composite entry for the tree children instead of an entry for each individual child. This entry is a reference to the map of children which already is stored in `TreeArtifactValue`. Adjust the lookup logic to work with composite entries in the map. Add a trie representation of the tree artifact locations to allow for faster response when searching for paths which do not exist in the map.

Change `ActionInputMap::getMetadata(String)` to `ActionInputMap::getMetadata(PathString)` so that we can pass the value from the callers (all of them already have it) and prevent creating garbage `PathFragment` when walking the trie.

PiperOrigin-RevId: 384297684