Cover `mobile-install` as part of the BuildTool `processRequest`.
Previously, we didn't have profile information available for the mobile-install portion of the command which did lead to confusing profiles and logging information.
This happened because as part of completing the request, in `stopRequest`, we did publish the `BuildComplete` event which in turn did cause the profile to stop and flush.

Also added a few other profile calls.

PiperOrigin-RevId: 384674626