Automated rollback of commit a95c7b7a9cb0088787d10a49d491b81f80cb6313.
*** Reason for rollback ***

Manual OOMs are not being triggered (b/193671557).

*** Original change description ***

Modify RetainedHeapLimiter to only throw an OOM when it itself triggers a manual GC. Other parts of Blaze can trigger manual GCs, and high memory usage after a manual GC does not mean that Blaze has to OOM.

Also don't trigger a manual GC if the previous GC was manual: this would lead to a double GC for any external manually triggered GCs.

PiperOrigin-RevId: 384763438