Eliminate NestedSet#forEachElement(). Introduce tests for NestedSetVisitor.
When NestedSetVisitor was first introduced it was implemented such that it could
only correctly visit a NestedSet with 'stable' order. It currently is implemented
identically to how `NestedSet#forEachElement()` is implemented and will visit a
NestedSet in the expected order. The JavaDoc has been updated and tests introduced
to verify this.

PiperOrigin-RevId: 384817322