Intern ParamFileInfo data objects.
ParamFileInfo is a data class that is used when registering Spawn actions. There is only a limited number of possible configurations for ParamFileInfo (flagFormatString is could have larger number of values but in practice only a couple of different values happen).

When registering actions natively this can be optimised in the rule implementation, while Starlark API does not provide an interface to do the optimisation.

PiperOrigin-RevId: 384873639