bzlmod: "fork" case for multiple_version_override
(https://github.com/bazelbuild/bazel/issues/13316)

It's okay for a module depends on foo@1.2 and foo@1.3 with different repo names as long as there's a multiple_version_override for foo that allows both 1.2 and 1.3. If only 1.3 is allowed or if there's no multiple_version_override at all, then selection will succeed right now, and the module will depend on foo@1.3 twice. This CL adds a check that this doesn't happen.

PiperOrigin-RevId: 384906574