Don't cache delegating DefaultInfo provider.
The cached value is rarely accessed again: it doesn't happen in native code. In starlark code it's possible to access it again - although it's rare (exploring current uses) and if it is a problem it could be optimised by reusing the value obtained in Starlark.

The change reduces retain memory - 24 bytes in every RuleConfiguredTarget (because of removed AtomicReference) and 24 bytes per each cached DelegatingDefaultInfo.

PiperOrigin-RevId: 385531451