Fix missing libc files in CcToolchainProvider.
When we removed middleman artifacts, CcToolchainProvider.getAllFilesMiddleman
was replaced with CcToolchainProvider.getAllFiles based on the existing
documentation of the methods. This one-for-one replacement
turned out to be insufficient, as the former also includes libc file (as an
implementation detail) and resulted in some builds failing because of missing
input files.

PiperOrigin-RevId: 385568176