Implement substitution in SpawnAction progress message.
The native SpawnAction provides formatter to improve performance (formatted string are not retained), while this was not available in Starlark.

The cl exposes formatting with default placeholders to Starlark and as well to native implementation. I marked old native api with formatting as deprecated as only the string with default placeholders can be used.

PiperOrigin-RevId: 385753127