bzlmod: Store canonical repo names in SelectionValue
(https://github.com/bazelbuild/bazel/issues/13316)

Up until now we assume the canonical repo name of a module is simply the module name. This doesn't work with multiple-version overrides. This CL addresses this issue; the canonical repo name will be ${moduleName}.${moduleVersion} for any module.

This does mean that we need to put in some extra work to build reverse lookup maps, and especially add extra logic in RepositoryMappingFunction to make sure that WORKSPACE references to repos generated by modules can still work (i.e. if B is a module, something that points to "@B" gets rewritten to "@B.1.0", the canonical repo name)

PiperOrigin-RevId: 385785118