Define BUILD_PROPELLER_TYPE="full" when propeller optimize is enabled.
Propeller build types support was added in unknown commit.

Further, this change also disables implicit split functions feature when
propeller optimize is used.  These two features are mutually exclusive and even
though the compiler will get this right, it is simpler to do this here.
Further, both features define propeller build type with different values, and
disabling implict split functions when propeller optimize is on makes sure the
build type is kept to the right value.  The one caveat to this is that users
can still explicitly turn on split functions and propeller optimize. There is
no real usage for this and we will not support it.

PiperOrigin-RevId: 385930318