Intern identifier and keyword Strings in the lexer
This makes it so that the lexer returns a (weakly) canonical String instance for all occurrences of a given identifier or keyword token.

This change was prompted by ilist@'s observation that there are many duplicate Strings that appear in profiling, consisting of common Starlark identifiers such as "name" and "ctx".

Benchmark results: Interning identifiers+keywords was observed to save 3-4% on a `bazel query deps(...)` invocation of bazel itself, and almost 1% on a much larger target. Static strong interning was a bit better than static weak interning, which in turn was a bit better than non-static (strong) interning. We went with static weak since static strong would constitute a memory leak for long-running Bazel servers; this follows precedent elsewhere in Bazel.

We considered interning bufferSlice() as well, which would intercept all conversions of the lexer's source text to String objects, but this showed no noticeable improvement and actually regressed RAM usage in some cases.

The parser already interns string literals; that is unaffected by this CL.

Note that we use Guava interners, not BlazeInterners, because this is Starlark interpreter code that must not depend on bazelisms.

RELNOTES: None
PiperOrigin-RevId: 386264085