Intern special case of "__pkg__" and "__subpackages__" labels
Motivation: ilist@ saw a lot of distinct "__pkg__" String objects show up in a profile.

We used to intern all values of Label's `name` field, but it was removed in unknown commit since it was counterproductive. Indeed, if I add back a static weak interner for `name`, I see +2.14% RAM on a bazel query deps() invocation over bazel itself. Discussion in that CL mentioned the possibility of doing a more advanced interning of popular names only (via something fancy like count-min sketch), but it was deemed not worth the extra complication. In contrast, interning __pkg__ / __subpackages__ seems like a freeby; it's easy to implement, and profiling showed a small but noticeable improvement for both of them.

Also update comments to point out that PackageIdentifier interning is weak and therefore doesn't guarantee a single canonical value.

PiperOrigin-RevId: 386316614