Fix a path name error in cc_common.link.
Currently, if cc_common.link is called for a library "foo/bar/libbaz.so", the output is "foo/bar/foo/bar/libbaz.so". This is because Blaze mishandles the subpath of the file during linking.

This change fixes this issue with doubled subdirectories, while preserving the intended effect of that path-handling code.

PiperOrigin-RevId: 386909211