Refactor apple_static_library use of split transition
The current way the native code is written cannot be re-written to Starlark. It looks like everything that cannot be done in Starlark can be refactored away. Mainly, the code used to pass the build configuration from each of the list of transitioned deps to lower level APIs that were using that configuration to for example build artifacts in a specific directory (which is unnecessary). In Starlark an action can only output to the directory of the current configured target.

RELNOTES:none
PiperOrigin-RevId: 387330041